+++ packages/notification-services-controller/src/NotificationServicesPushController/services/services.ts
@@ -1,13 +1,9 @@
-import log from 'loglevel';
-
 import * as endpoints from './endpoints';
-import type { CreateRegToken, DeleteRegToken } from './push';
-import {
-  listenToPushNotificationsClicked,
-  listenToPushNotificationsReceived,
-} from './push/push-web';
-import type { Types } from '../../NotificationServicesController';
 import type { PushNotificationEnv } from '../types';
+import type {
+  CreateRegToken,
+  DeleteRegToken,
+} from '../types/push-service-interface';
 
 export type RegToken = {
   token: string;
@@ -19,33 +15,8 @@ export type RegToken = {
  */
 export type LinksResult = {
   trigger_ids: string[];
-
   registration_tokens: RegToken[];
 };
-
-/**
- * Fetches push notification links from a remote endpoint using a BearerToken for authorization.
- *
- * @param bearerToken - The JSON Web Token used for authorization.
- * @returns A promise that resolves with the links result or null if an error occurs.
- */
-export async function getPushNotificationLinks(
-  bearerToken: string,
-): Promise<LinksResult | null> {
-  try {
-    const response = await fetch(endpoints.REGISTRATION_TOKENS_ENDPOINT, {
-      headers: { Authorization: `Bearer ${bearerToken}` },
-    });
-    if (!response.ok) {
-      log.error('Failed to fetch the push notification links');
-      throw new Error('Failed to fetch the push notification links');
-    }
-    return response.json() as Promise<LinksResult>;
-  } catch (error) {
-    log.error('Failed to fetch the push notification links', error);
-    return null;
-  }
-}
 
 /**
  * Updates the push notification links on a remote API.
@@ -88,7 +59,6 @@ type ActivatePushNotificationsParams = {
   env: PushNotificationEnv;
   createRegToken: CreateRegToken;
   platform: 'extension' | 'mobile' | 'portfolio';
-  fcmToken?: string;
 };
 
 /**
@@ -100,32 +70,20 @@ type ActivatePushNotificationsParams = {
 export async function activatePushNotifications(
   params: ActivatePushNotificationsParams,
 ): Promise<string | null> {
-  const { bearerToken, triggers, env, createRegToken, platform, fcmToken } =
-    params;
+  const { bearerToken, triggers, env, createRegToken, platform } = params;
 
-  const notificationLinks = await getPushNotificationLinks(bearerToken);
-
-  if (!notificationLinks) {
-    return null;
-  }
-
-  const regToken = fcmToken ?? (await createRegToken(env).catch(() => null));
+  const regToken = await createRegToken(env).catch(() => null);
   if (!regToken) {
     return null;
   }
 
-  const newRegTokens = new Set(notificationLinks.registration_tokens);
-  newRegTokens.add({ token: regToken, platform });
-
-  await updateLinksAPI(bearerToken, triggers, Array.from(newRegTokens));
+  await updateLinksAPI(bearerToken, triggers, [{ token: regToken, platform }]);
   return regToken;
 }
 
 type DeactivatePushNotificationsParams = {
   // Push Links
   regToken: string;
-  bearerToken: string;
-  triggers: string[];
 
   // Push Un-registration
   env: PushNotificationEnv;
@@ -133,37 +91,20 @@ type DeactivatePushNotificationsParams = {
 };
 
 /**
- * Disables push notifications by removing the registration token and unlinking triggers.
+ * Disables push notifications by removing the registration token
+ * We do not need to unlink triggers, and remove old reg tokens (this is cleaned up in the back-end)
  *
  * @param params - Deactivate Push Params
- * @returns A promise that resolves with true if notifications were successfully disabled, false otherwise.
+ * @returns A promise that resolves with true if push notifications were successfully disabled, false otherwise.
  */
 export async function deactivatePushNotifications(
   params: DeactivatePushNotificationsParams,
 ): Promise<boolean> {
-  const { regToken, bearerToken, triggers, env, deleteRegToken } = params;
+  const { regToken, env, deleteRegToken } = params;
 
   // if we don't have a reg token, then we can early return
   if (!regToken) {
     return true;
-  }
-
-  const notificationLinks = await getPushNotificationLinks(bearerToken);
-  if (!notificationLinks) {
-    return false;
-  }
-
-  const filteredRegTokens = notificationLinks.registration_tokens.filter(
-    (r) => r.token !== regToken,
-  );
-
-  const isTokenRemovedFromAPI = await updateLinksAPI(
-    bearerToken,
-    triggers,
-    filteredRegTokens,
-  );
-  if (!isTokenRemovedFromAPI) {
-    return false;
   }
 
   const isTokenRemovedFromFCM = await deleteRegToken(env);
@@ -176,7 +117,6 @@ export async function deactivatePushNotifications(
 
 type UpdateTriggerPushNotificationsParams = {
   // Push Links
-  regToken: string;
   bearerToken: string;
   triggers: string[];
 
@@ -202,12 +142,10 @@ type UpdateTriggerPushNotificationsParams = {
 export async function updateTriggerPushNotifications(
   params: UpdateTriggerPushNotificationsParams,
 ): Promise<{
-  isTriggersLinkedToPushNotifications: boolean;
-  fcmToken?: string | null;
+  fcmToken: string;
 }> {
   const {
     bearerToken,
-    regToken,
     triggers,
     createRegToken,
     platform,
@@ -215,79 +153,20 @@ export async function updateTriggerPushNotifications(
     env,
   } = params;
 
-  const notificationLinks = await getPushNotificationLinks(bearerToken);
-  if (!notificationLinks) {
-    return { isTriggersLinkedToPushNotifications: false };
-  }
-  // Create new registration token if doesn't exist
-  const hasRegToken = Boolean(
-    regToken &&
-      notificationLinks.registration_tokens.some((r) => r.token === regToken),
-  );
-
-  let newRegToken: string | null = null;
-  if (!hasRegToken) {
-    await deleteRegToken(env);
-    newRegToken = await createRegToken(env);
-    if (!newRegToken) {
-      throw new Error('Failed to create a new registration token');
-    }
-    notificationLinks.registration_tokens.push({
-      token: newRegToken,
-      platform,
-    });
+  await deleteRegToken(env);
+  const newRegToken = await createRegToken(env);
+  if (!newRegToken) {
+    throw new Error('Failed to create a new registration token');
   }
 
-  const isTriggersLinkedToPushNotifications = await updateLinksAPI(
-    bearerToken,
-    triggers,
-    notificationLinks.registration_tokens,
-  );
+  const linksNotUpdated = await updateLinksAPI(bearerToken, triggers, [
+    { token: newRegToken, platform },
+  ]);
+  if (!linksNotUpdated) {
+    throw new Error('Failed to create links to new reg token');
+  }
 
   return {
-    isTriggersLinkedToPushNotifications,
-    fcmToken: newRegToken ?? null,
+    fcmToken: newRegToken,
   };
-}
-
-type ListenToPushNotificationsParams = {
-  env: PushNotificationEnv;
-  listenToPushReceived: (
-    notification: Types.INotification,
-  ) => void | Promise<void>;
-  listenToPushClicked: (
-    event: NotificationEvent,
-    notification?: Types.INotification,
-  ) => void;
-};
-
-/**
- * Listens to push notifications and invokes the provided callback function with the received notification data.
- *
- * @param params - listen params
- * @returns A promise that resolves to an unsubscribe function to stop listening to push notifications.
- */
-export async function listenToPushNotifications(
-  params: ListenToPushNotificationsParams,
-): Promise<() => void> {
-  const { env, listenToPushReceived, listenToPushClicked } = params;
-
-  /*
-  Push notifications require 2 listeners that need tracking (when creating and for tearing down):
-  1. handling receiving a push notification (and the content we want to display)
-  2. handling when a user clicks on a push notification
-  */
-  const unsubscribePushNotifications = await listenToPushNotificationsReceived(
-    env,
-    listenToPushReceived,
-  );
-  const unsubscribeNotificationClicks =
-    listenToPushNotificationsClicked(listenToPushClicked);
-
-  const unsubscribe = () => {
-    unsubscribePushNotifications?.();
-    unsubscribeNotificationClicks();
-  };
-
-  return unsubscribe;
 }